requirements.txt
pymongo
flask
uwsgi
flask-cors
pyjwt
